Removes password component and moves eye icon to registration form
