--- web/beacon-app/src/features/auth/components/Password/index.tsx
@@ -1,1 +1,0 @@
-export { default as Password } from './Password';
