+++ web/beacon-app/src/features/auth/components/Register/__tests__/RegistrationForm.spec.tsx
@@ -33,6 +33,44 @@ describe('first', () => {
     });
   });
 
+  describe('Password eye icon renders', () => {
+    it('should show icon with a closed eye, hidden password text, and accessible button name "Show Password" on render', () => {
+      const handleSubmit = vi.fn();
+      render(<RegistrationForm onSubmit={handleSubmit} />);
+      const button = screen.getByTestId('button');
+
+      expect(screen.getByTestId('hidePassword')).toBeVisible;
+      expect(screen.getByTestId('password')).toHaveAttribute('type', 'password');
+      expect(button).toHaveAccessibleName('Show Password');
+    });
+  });
+
+  describe('Show password', () => {
+    it('should show icon with open eye, password text, and accessible button name "Hide Password" when clicks on icon', () => {
+      const handleSubmit = vi.fn();
+      render(<RegistrationForm onSubmit={handleSubmit} />);
+      const button = screen.getByTestId('button');
+
+      fireEvent.click(button);
+      expect(screen.getByTestId('showPassword')).toBeVisible;
+      expect(screen.getByTestId('password')).toHaveAttribute('type', 'text');
+      expect(button).toHaveAccessibleName('Hide Password');
+    });
+  });
+
+  describe('Hide password', () => {
+    it('should show icon with eye closed, hidden password text, and accessible button name "Show Password" when user clicks on eye icon a 2nd time', () => {
+      const handleSubmit = vi.fn();
+      render(<RegistrationForm onSubmit={handleSubmit} />);
+      const button = screen.getByTestId('button');
+
+      fireEvent.doubleClick(button);
+      expect(screen.getByTestId('hidePassword')).toBeVisible;
+      expect(screen.getByTestId('password')).toHaveAttribute('type', 'password');
+      expect(button).toHaveAccessibleName('Show Password');
+    });
+  });
+
   // describe('Password', () => {
   //   it('should display error message when password field is empty', async () => {
   //     const handleSubmit = vi.fn();
